Drivers License App to check the number of correct and incorrect answers of the exam, and predict the pass/fail status of the exam.
Application using TabControl and Listbox to read the student’s answers for each of the 20 questions from a text file and store the answers in another array. After the student’s answers have been read from the file, the program displays a message indicating whether the student passed or failed the exam. A student must correctly answer 15 questions to pass. It then displays the total number of correctly answered questions, the total number of incorrectly answered questions, and a list showing the question numbers of the incorrectly answered questions.